+++ fit-frontend/src/Fit/Start/BasicInfo.jsx
